[refactor] MemoizeRecords -> MemoizeQuery
