--- lib/redis_memo/memoize_method.rbi
@@ -1,10 +1,0 @@
-# typed: true
-
-module RedisMemo::MemoizeMethod
-  include Kernel
-
-  def class; end
-  def name; end
-  def alias_method(*); end
-  def define_method(*); end
-end
