+++ lib/redis_memo/memoize_query/memoize_table_column.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+require_relative 'memoize_method'
+
+module RedisMemo::MemoizeQuery
+end
